finally got everything wired and updated filetree
