+++ filetree.txt
@@ -1,13 +1,33 @@
 My current file tree
 
 
-Quote-generator/
-├─ css/
-│  ├─ main.css/
-├─ data/
-│  ├─ responses.js
-├─ javascript/
-│  ├─ main.js
-├─ index.html
-├─ .gitattributes
-├─ README.md
+/quote-generator
+│
+├── /css
+│   ├── /components
+│   │   ├── header.css
+│   │   ├── welcome.css
+│   │   ├── card-grid.css
+│   │   ├── quote-submit.css
+│   │   ├── pending-quotes.css
+│   │   └── buttons.css
+│   ├── globals.css
+│   └── main.css
+│
+├── /javascript
+│   ├── /components
+│   │   ├── header.js
+│   │   ├── welcome.js
+│   │   ├── card-grid.js
+│   │   ├── quote-submit.js
+│   │   └── pending-quotes.js
+│   ├── state.js
+│   ├── utils.js
+│   ├── brokethings.js
+│   └── main.js
+│
+├── /data
+│   ├── imagesData.js
+│   └── responses.js
+│
+└── index.html
