+++ javascript/components/header.js
@@ -1,6 +1,6 @@
 import { state } from "../state.js";
 
-export const handleNavigation = (event) => {
+/*export const handleNavigation = (event) => {
   const action = event.target.dataset.action;
   if (action === "open-submit") {
     state.currentView = "quote-submit";
@@ -13,7 +13,28 @@ export const handleNavigation = (event) => {
     }
     toggleView(state.currentView);
   }
-}; 
+};*/ 
+
+export const handleNavigation = (event) => {
+  const action = event.target.dataset.action;
+  if (action === 'open-submit') {
+    state.previousView = state.currentView;
+    state.currentView = 'quote-submit';
+    toggleView('quote-submit');
+  } else if (action === 'close-submit') {
+    // Handle "Back to Start" button
+    if (state.currentView === 'quote-submit') {
+      state.currentView = state.previousView || 'welcome'; // Fallback to 'welcome' if no previous view
+    } else if (state.currentView === 'card-grid') {
+      state.currentView = 'welcome'; // Go back to welcome from card grid
+    }
+    toggleView(state.currentView);
+  } else if (action === 'go-back') {
+    // Handle "Back to Start" button
+    state.currentView = 'welcome'; // Always go back to welcome
+    toggleView('welcome');
+  }
+};
 
 export const toggleView = (view) => {
   const views = ["welcome", "card-grid", "quote-submit"];
